refactor: setup-gh実行時にGitリポジトリ情報を自動取得
- StartSession時のGITHUB_REPO_FULLNAME環境変数設定を削除
- setup-gh実行時にgetGitHubRepoFullName()でリポジトリ情報を自動取得
- SSH、HTTPS、gitプロトコルのURL形式に対応
- 認証不要でgit config から自動検出
